% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_creation.R
\name{add_external_seq_pq}
\alias{add_external_seq_pq}
\title{Add external sequences to a phyloseq object}
\usage{
add_external_seq_pq(physeq, ext_seqs, prefix = "external_")
}
\arguments{
\item{physeq}{(required): a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{ext_seqs}{A DNAStringSet object}

\item{prefix}{A prefix add to the taxa name}
}
\value{
A phyloseq object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Useful to compute true negative values with functions
\code{\link[=tc_metrics_mock]{tc_metrics_mock()}} and \code{\link[=tc_metrics_mock_vec]{tc_metrics_mock_vec()}}.

Note the that the tax_table for additional sequences is full of NA and
that the corresponding otu_table is full of 0.
}
\examples{
add_external_seq_pq(
  data_fungi_mini,
  Biostrings::readDNAStringSet(
    system.file("extdata/ex_little.fasta", package = "MiscMetabar")
  )
)

}
\author{
Adrien Taudière
}
