% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_phyloseq.R
\name{add_phyloseq}
\alias{add_phyloseq}
\title{Add a phyloseq object to a list_phyloseq}
\usage{
add_phyloseq(x, physeq, name = NULL)
}
\arguments{
\item{x}{(required) A list_phyloseq object.}

\item{physeq}{(required) A phyloseq object to add.}

\item{name}{(character, default NULL) Optional name for the new phyloseq object.
If NULL, a name is generated automatically.}
}
\value{
A new list_phyloseq object with the added phyloseq
}
\description{
Add a phyloseq object to a list_phyloseq
}
