% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_lpq.R
\name{analysis_lpq}
\alias{analysis_lpq}
\title{Analysis functions for list_phyloseq objects}
\description{
Functions to run statistical analyses on each phyloseq object
in a list_phyloseq and return summary tables.
}
