% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_lpq.R
\name{ancombc_lpq}
\alias{ancombc_lpq}
\title{ANCOM-BC analysis on each phyloseq object in a list_phyloseq}
\usage{
ancombc_lpq(x, fact, levels_fact = NULL, tax_level = NULL, verbose = TRUE, ...)
}
\arguments{
\item{x}{(required) A list_phyloseq object.}

\item{fact}{(character, required) The name of a column in \code{sample_data}
to use as the grouping factor. Must be present in \strong{all} phyloseq
objects.}

\item{levels_fact}{(character vector, default NULL) Levels of the factor
to include. If NULL, all levels are used.}

\item{tax_level}{(character, default "Class") Taxonomic level for
agglomeration before analysis.}

\item{verbose}{(logical, default TRUE) If TRUE, print progress messages.}

\item{...}{Additional arguments passed to \code{\link[MiscMetabar:ancombc_pq]{MiscMetabar::ancombc_pq()}}.}
}
\value{
A tibble with the combined ANCOM-BC results from all phyloseq
objects. Contains the \verb{$res} data frame from each ANCOM-BC run, with
an additional \code{name} column identifying the source phyloseq object.
}
\description{
Performs ANCOM-BC differential abundance analysis using
\code{\link[MiscMetabar:ancombc_pq]{MiscMetabar::ancombc_pq()}} on each phyloseq object in a list_phyloseq
and returns a combined result table.
}
\details{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


This function requires that the list_phyloseq type is NOT
\code{SEPARATE_ANALYSIS}, as the factor must be common across all phyloseq
objects.

When no common taxa exist across the phyloseq objects, taxa names are
suffixed with the phyloseq object name to make them distinguishable.
}
\examples{
data_fungi_high <- multiply_counts_pq(data_fungi, "Height", "High", 2)

lpq <- list_phyloseq(
  list(
    fungi = data_fungi,
    fungi_height = data_fungi_high
  )
)

results <- ancombc_lpq(lpq, fact = "Height")
results

results |>
  filter(diff_HeightLow) |>
  ggplot(aes(y = taxon, x = lfc_HeightLow, color = name, shape = q_HeightLow < 0.1)) +
  geom_point(size = 4) +
  geom_vline(xintercept = 0) +
  theme_minimal()

results_Genus <- ancombc_lpq(lpq, fact = "Height", tax_level = "Genus")

results_Genus |>
  filter(diff_HeightLow) |>
  ggplot(aes(y = taxon, x = lfc_HeightLow, color = name, shape = q_HeightLow < 0.1)) +
  geom_point(size = 4) +
  geom_vline(xintercept = 0) +
  theme_minimal()

}
\seealso{
\code{\link[MiscMetabar:ancombc_pq]{MiscMetabar::ancombc_pq()}}, \code{\link[=aldex_lpq]{aldex_lpq()}}, \code{\link[=multipatt_lpq]{multipatt_lpq()}}
}
