% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles_pq.R
\name{bubbles_pq}
\alias{bubbles_pq}
\title{Bubble plot of phyloseq object with observablehq}
\usage{
bubbles_pq(
  physeq,
  rank_label = "Taxa",
  rank_color = "Family",
  categorical_scheme = "d3.schemeCategory10",
  label_color = "grey10",
  value_color = "grey20",
  label_size = 8,
  log1ptransform = FALSE,
  min_nb_seq = 0,
  randomize = FALSE,
  seed = 32,
  width = 600,
  include = c("TitleCell", "key", "chart"),
  notebook = "https://observablehq.com/d/d755af3197af2320",
  return_dataframe = FALSE,
  title = "",
  title_size = "22px"
)
}
\arguments{
\item{physeq}{(required): a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{rank_label}{The name of the column in the @tax_table slot to label the
points. If set to "Taxa" (default) the taxa name is used}

\item{rank_color}{The name of the column in the @tax_table slot to color the
points.}

\item{categorical_scheme}{A color scheme from \href{https://d3js.org/d3-scale-chromatic/categorical}{d3js}}

\item{label_color}{Color for the label text.}

\item{value_color}{Color for the value text.}

\item{label_size}{Font size for the label.}

\item{log1ptransform}{(logical, default TRUE) If TRUE,
the number of sequences is log1p transformed.}

\item{min_nb_seq}{Minimum number of sequences to filter out taxa with low
abundance}

\item{randomize}{If TRUE, shuffles the order of the taxa. Default FALSE.}

\item{seed}{(int) Set seed for the randomization.}

\item{width}{The notebook width for visualization in pixels. Default 600.}

\item{include}{You can modify include to remove the title and/or the legend
for ex. using \code{include = c("chart")}}

\item{notebook}{You can change the notebook url if you know what you are
doing.}

\item{return_dataframe}{(logical, default FALSE) If TRUE, the plot is
not returned, but the resulting dataframe to plot is returned.}

\item{title}{Title of the plot}

\item{title_size}{Font size for the title}
}
\value{
A htmlwidget
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}
}
\examples{
bubbles_pq(physeq = data_fungi, rank_color = "Class", randomize = TRUE)
bubbles_pq(physeq = data_fungi, min_nb_seq = 10000, categorical_scheme = "d3.schemePastel1")
bubbles_pq(physeq = data_fungi, min_nb_seq = 1000, categorical_scheme = "d3.schemeTableau10", label_color = "purple", value_color = "white")
bubbles_pq(physeq = data_fungi, rank_color = "Order", include = c("chart"), log1ptransform = TRUE)
bubbles_pq(physeq = data_fungi, rank_label = "Genus", rank_color = "Class", include = c("chart"), randomize = TRUE, seed = 3)
bubbles_pq(physeq = as_binary_otu_table(data_fungi), min_nb_seq = 20, title = "Nb of samples per OTU. <br> Only OTU present in more than 20 samples are shown", title_size = "12px")
}
\author{
Adrien Taudière
}
