% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_progress.R
\name{tc_df_pq}
\alias{tc_df_pq}
\title{Cross-tabulation of two taxonomic ranks}
\usage{
tc_df_pq(physeq, rank_1 = "Family", rank_2 = "Class", ...)
}
\arguments{
\item{physeq}{(required) A \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{rank_1}{(character, default "Family") The name of the first taxonomic
rank (column in tax_table slot) for the cross-tabulation rows.}

\item{rank_2}{(character, default "Class") The name of the second taxonomic
rank (column in tax_table slot) for the cross-tabulation columns.}

\item{...}{Additional arguments passed to \code{\link[gtsummary:tbl_cross]{gtsummary::tbl_cross()}}.}
}
\value{
A gtsummary tbl_cross object displaying the cross-tabulation of
the two taxonomic ranks.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Creates a cross-tabulation (contingency table) comparing two taxonomic ranks
from a phyloseq object. Useful for comparing taxonomic assignments from
different databases, algorithms, or taxonomic levels.
}
\examples{
\dontrun{
tc_df_pq(data_fungi_mini)
tc_df_pq(data_fungi_mini, rank_1 = "Order", rank_2 = "Family")

# Compare taxonomic assignments from different methods
ref_fasta <- system.file("extdata", "mini_UNITE_fungi.fasta.gz",
  package = "MiscMetabar", mustWork = TRUE
)
data_fungi_mini2 <- data_fungi_mini |>
  add_new_taxonomy_pq(ref_fasta, suffix = "_sintax", method = "sintax") |>
  add_new_taxonomy_pq(ref_fasta, suffix = "_lca", method = "lca")

tc_df_pq(data_fungi_mini2, rank_1 = "Class_lca", rank_2 = "Class_sintax")
}
}
\author{
Adrien Taudière
}
