% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_phyloseq.R
\name{length.list_phyloseq}
\alias{length.list_phyloseq}
\title{Number of phyloseq objects in a \code{list_phyloseq}}
\usage{
\method{length}{list_phyloseq}(x)
}
\arguments{
\item{x}{A \code{list_phyloseq} object.}
}
\value{
An integer: the number of phyloseq objects stored.
}
\description{
Number of phyloseq objects in a \code{list_phyloseq}
}
\examples{
lpq <- list_phyloseq(list(a = data_fungi, b = data_fungi))
length(lpq) # 2
}
