% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_phyloseq.R
\name{names.list_phyloseq}
\alias{names.list_phyloseq}
\title{Names of phyloseq objects in a \code{list_phyloseq}}
\usage{
\method{names}{list_phyloseq}(x)
}
\arguments{
\item{x}{A \code{list_phyloseq} object.}
}
\value{
A character vector of names.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}
}
\examples{
lpq <- list_phyloseq(list(a = data_fungi, b = data_fungi))
names(lpq) # c("a", "b")
}
