% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_lpq.R
\name{print.adonis_lpq_result}
\alias{print.adonis_lpq_result}
\title{Print method for adonis_lpq results}
\usage{
\method{print}{adonis_lpq_result}(x, ...)
}
\arguments{
\item{x}{An adonis_lpq_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for adonis_lpq results
}
