% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_lpq.R
\name{print.estim_diff_lpq_result}
\alias{print.estim_diff_lpq_result}
\title{Print categorical estimation statistics for a list of phyloseq objects}
\usage{
\method{print}{estim_diff_lpq_result}(x, ...)
}
\arguments{
\item{x}{An \code{estim_diff_lpq_result} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print categorical estimation statistics for a list of phyloseq objects.
\if{html}{\out{<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">}}
\if{html}{\out{<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental">}}\if{html}{\out{</a>}}
}
\details{
Print categorical estimation statistics for a list of phyloseq objects.
}
