% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_phyloseq.R
\name{print.list_phyloseq}
\alias{print.list_phyloseq}
\title{Print a \code{list_phyloseq} object}
\usage{
\method{print}{list_phyloseq}(x, ...)
}
\arguments{
\item{x}{A \code{list_phyloseq} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{list_phyloseq} object
}
