% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxo_plot.R
\name{rainplot_taxo_na}
\alias{rainplot_taxo_na}
\title{Rainplot of the nb taxa assigned (not NA)}
\usage{
rainplot_taxo_na(
  physeq,
  ranks = NULL,
  min_nb_seq = 0,
  merge_sample_by = NULL,
  sample_colored = FALSE,
  sample_linked = FALSE,
  ...
)
}
\arguments{
\item{physeq}{(required): a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{ranks}{The ranks to include in the rainplot. If left to NULL, all
ranks are used. Each rank can be defined either by integer for the index
or by its full names (exacly matching the colnames of the \code{tax_table} slot).}

\item{min_nb_seq}{Minimum number of sequences to filter out taxa with low
abundance}

\item{merge_sample_by}{a vector to determine
which samples to merge using \code{\link[MiscMetabar:merge_samples2]{MiscMetabar::merge_samples2()}} function.
Need to be in \code{physeq@sam_data}}

\item{sample_colored}{(logical, default FALSE) Do points are colored by
samples?}

\item{sample_linked}{(logical, default FALSE) Do points are linked by
samples?}

\item{...}{}
}
\value{
A ggplot2 object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Useful to compare taxonomy from two different source/db/algo
}
\examples{
rainplot_taxo_na(subset_taxa_pq(Glom_otu, taxa_sums(Glom_otu) > 5000),
  ranks = c("Family", "Family__eukaryome_Glomero")
)
\dontrun{
rainplot_taxo_na(Glom_otu)

Glom_otu@sam_data$tmt_type <- paste0(Glom_otu@sam_data$Tmt, "_", Glom_otu@sam_data$Type)
rainplot_taxo_na(
  Glom_otu,
  merge_sample_by = "tmt_type",
  sample_colored = TRUE,
  sample_linked = TRUE
)
rainplot_taxo_na(Glom_otu, ranks = c(4, 12), rain.side = "f1x1")
rainplot_taxo_na(
  Glom_otu,
  ranks = c(6, 14),
  rain.side = "f1x1",
  sample_linked = TRUE
) +
  theme(legend.position = "none")
}
}
\author{
Adrien Taudière
}
