% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtab_modification.R
\name{rename_ranks_pq}
\alias{rename_ranks_pq}
\title{Rename names of ranks in the tax_table slot of a phyloseq object}
\usage{
rename_ranks_pq(
  physeq,
  old_names = NULL,
  new_names = NULL,
  pattern = NULL,
  replacement = NULL,
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{old_names}{Names of the names to replace}

\item{new_names}{Names of the new names}

\item{pattern}{Pattern to replace by the args replacement}

\item{replacement}{Pattern of replacement}

\item{fixed}{see ?grep}

\item{perl}{see ?grep}

\item{useBytes}{see ?grep}
}
\value{
An object of class phyloseq
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


The users can use the function using a couple of vector \code{old_names}/\code{new_names}
or using a \code{pattern} to replace by \code{replacement}
}
\examples{
rename_ranks_pq(data_fungi, c("Confidence.Ranking", "Phylum"), c("Conf.Rank.Guild", "Phyla"))@tax_table |>
  colnames()
rename_ranks_pq(data_fungi, pattern = ".", replacement = "_", fixed = TRUE)@tax_table |>
  colnames()
}
\author{
Adrien Taudière
}
