% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtab_modification.R
\name{select_ranks_pq}
\alias{select_ranks_pq}
\title{Select taxonomic ranks in a phyloseq object}
\usage{
select_ranks_pq(physeq, ...)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{...}{\if{html}{\out{<tidy-select>}} One or more unquoted expressions separated by commas. Variable names can be used as if they were positions in the data frame, so expressions like x:y can be used to select a range of variables. See ?dplyr::select}
}
\value{
A phyloseq object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Useful in tidyverse-like pipeline.
}
\examples{
select_ranks_pq(data_fungi, Order, Family)@tax_table |>
  dim()
select_ranks_pq(data_fungi, !Order)@tax_table |>
  colnames()
#
}
\author{
Adrien Taudière
}
