% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable_lpq.R, R/list_phyloseq.R
\name{shared_mod_lpq}
\alias{shared_mod_lpq}
\title{Display shared sample_data modalities}
\usage{
shared_mod_lpq(x, max_modalities = NULL)

shared_mod_lpq(x, max_modalities = NULL)
}
\arguments{
\item{x}{(required) A list_phyloseq object.}

\item{max_modalities}{(integer, default 10) Maximum number of modalities to
display per variable.}
}
\value{
A tibble or NULL if no shared modalities exist

A tibble or NULL if no shared modalities exist
}
\description{
Create a table showing which sample_data variable modalities
are shared across phyloseq objects in a list_phyloseq.

Create a table showing which sample_data variable modalities
are shared across phyloseq objects in a list_phyloseq.
}
\details{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}

\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}
}
\examples{
lpq <- list_phyloseq(list(data1 = data_fungi, data2 = data_fungi_mini))

shared_mod_lpq(lpq)
shared_mod_lpq(lpq, 10)
lpq <- list_phyloseq(list(data1 = data_fungi, data2 = data_fungi_mini))

shared_mod_lpq(lpq)
shared_mod_lpq(lpq, 10)
}
