% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_phyloseq.R
\name{[.list_phyloseq}
\alias{[.list_phyloseq}
\title{Subset a \code{list_phyloseq} object}
\usage{
\method{[}{list_phyloseq}(x, i)
}
\arguments{
\item{x}{A \code{list_phyloseq} object.}

\item{i}{Index (integer or character) selecting elements.}
}
\value{
A new \code{list_phyloseq} built from the selected phyloseq objects.
}
\description{
Subset a \code{list_phyloseq} object
}
\examples{
lpq <- list_phyloseq(list(a = data_fungi, b = data_fungi))
lpq[1]
}
