% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_phyloseq.R
\name{[[.list_phyloseq}
\alias{[[.list_phyloseq}
\title{Extract a single phyloseq object from a \code{list_phyloseq}}
\usage{
\method{[[}{list_phyloseq}(x, i)
}
\arguments{
\item{x}{A \code{list_phyloseq} object.}

\item{i}{Index (integer or character) selecting one element.}
}
\value{
The selected \code{phyloseq} object.
}
\description{
Extract a single phyloseq object from a \code{list_phyloseq}
}
\examples{
lpq <- list_phyloseq(list(a = data_fungi, b = data_fungi))
lpq[[1]]
}
