% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxo_plot.R
\name{tc_circle}
\alias{tc_circle}
\title{Circle of correspondence between two taxonomic levels}
\usage{
tc_circle(
  physeq,
  rank_1 = NULL,
  rank_2 = NULL,
  suffix_1 = "_1",
  suffix_2 = "_2"
)
}
\arguments{
\item{physeq}{(required): a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{rank_1}{Define the first taxonomic rank as the number or the name
of the column in tax_table slot}

\item{rank_2}{Define the second taxonomic rank as the number or the name
of the column in tax_table slot}

\item{suffix_1}{A suffix to add to rank_1 values (default "_1")}

\item{suffix_2}{A suffix to add to rank_2 values (default "_2")}
}
\value{
A ggplot2 object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Useful to compare taxonomy from two different source/db/algo side-by-side
}
\examples{
tc_circle(
  Glom_otu,
  "Genus__eukaryome_Glomero",
  "Genus",
  suffix_1 = "_Euk",
  suffix_2 = "_Marjaam"
)

tc_circle(
  Glom_otu,
  "Family__eukaryome_Glomero",
  "Family",
  suffix_1 = "_Euk",
  suffix_2 = "_Marjaam"
)
}
\author{
Adrien Taudière
}
