% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxo_plot.R
\name{tc_points_matrix}
\alias{tc_points_matrix}
\title{Matrix of point to compare two taxonomic ranks}
\usage{
tc_points_matrix(
  physeq,
  rank_1,
  rank_2,
  color_1 = "#dc863b",
  color_2 = "#2e7891",
  stat_across_sample = "sum",
  merge_sample_by = NULL
)
}
\arguments{
\item{physeq}{(required): a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{rank_1}{Define the first taxonomic rank as the number or the name
of the column in tax_table slot}

\item{rank_2}{Define the second taxonomic rank as the number or the name
of the column in tax_table slot}

\item{color_1}{Color for rank_1 values}

\item{color_2}{Color for rank_1 values}

\item{stat_across_sample}{Either "mean" or "sum". Set how the abundance is
computed across samples.}

\item{merge_sample_by}{a vector to determine
which samples to merge using \code{\link[MiscMetabar:merge_samples2]{MiscMetabar::merge_samples2()}} function.
Need to be in \code{physeq@sam_data}}
}
\value{
A ggplot2 object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Useful to compare taxonomy from two different source/db/algo side-by-side
}
\examples{

tc_points_matrix(
  subset_taxa_pq(Glom_otu, taxa_sums(Glom_otu) > 5000),
  "Order", "Order__eukaryome_Glomero"
)
\dontrun{
tc_points_matrix(Glom_otu, 6, 14)
tc_points_matrix(Glom_otu, 4, 12)
tc_points_matrix(Glom_otu, 4, 12, stat_across_sample = "mean")

Glom_otu@sam_data$unique_value <- rep("samp", nsamples(Glom_otu))
tc_points_matrix(as_binary_otu_table(Glom_otu), 5, 13,
  stat_across_sample = "sum", merge_sample_by = "unique_value"
)
tc_points_matrix(as_binary_otu_table(Glom_otu), 5, 13,
  stat_across_sample = "mean"
)
tc_points_matrix(Glom_otu, 5, 13,
  stat_across_sample = "mean"
)
}
}
\author{
Adrien Taudière
}
