% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_sankey.R
\name{tc_sankey}
\alias{tc_sankey}
\title{Sankey diagram to compare two taxonomic ranks}
\usage{
tc_sankey(physeq, rank_1, rank_2, fill_by = "rank_1")
}
\arguments{
\item{physeq}{(required) A \code{\link[phyloseq]{phyloseq-class}} object
obtained using the \code{phyloseq} package.}

\item{rank_1}{(character or integer, required) Define the first taxonomic
rank as the number or the name of the column in tax_table slot.}

\item{rank_2}{(character or integer, required) Define the second taxonomic
rank as the number or the name of the column in tax_table slot.}

\item{fill_by}{(character, default "rank_1") Which rank to use for fill
color. Either "rank_1" or "rank_2".}
}
\value{
A ggplot2 object that can be further customized.
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Creates a Sankey (alluvial) diagram to visualize the correspondence between
two taxonomic ranks. Useful for comparing taxonomy assignments from different
databases or algorithms.
}
\examples{
tc_sankey(
  Glom_otu,
  "Class__eukaryome_Glomero",
  "Class"
)

# Fill by rank_2 instead
tc_sankey(
  Glom_otu,
  "Class__eukaryome_Glomero",
  "Class",
  fill_by = "rank_2"
)

\dontrun{
# Add labels to the strata
tc_sankey(Glom_otu, "Class__eukaryome_Glomero", "Class") +
  geom_label(
    stat = "stratum",
    aes(label = after_stat(stratum)),
    na.rm = TRUE
  ) +
  theme(legend.position = "none")
}
}
\author{
Adrien Taudière
}
