% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_creation.R
\name{add_shuffle_seq_pq}
\alias{add_shuffle_seq_pq}
\title{Add fake sequences by shuffling existing ones in a phyloseq object}
\usage{
add_shuffle_seq_pq(physeq, n_fake = NULL, prop_fake = NULL, prefix = "fake_")
}
\arguments{
\item{physeq}{(required): a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{n_fake}{A number of fake taxa added.}

\item{prop_fake}{A proportion of fake taxa added.}

\item{prefix}{A prefix add to the taxa name}
}
\value{
A phyloseq object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Useful to compute true negative values with functions
\code{\link[=tc_metrics_mock]{tc_metrics_mock()}} and \code{\link[=tc_metrics_mock_vec]{tc_metrics_mock_vec()}}.

Note the that the tax_table for additional sequences is full of NA and
that the corresponding otu_table is full of 0.
}
\examples{
d_fake_F <- data_fungi_mini |>
  add_shuffle_seq_pq(prop_fake = 0.1)
ntaxa(d_fake_F) - ntaxa(data_fungi_mini)
}
\author{
Adrien Taudière
}
