% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattable_lpq.R
\name{factor_formatter}
\alias{factor_formatter}
\title{Format factor columns with funky colored backgrounds}
\usage{
factor_formatter(x)
}
\arguments{
\item{x}{A factor or character vector}
}
\value{
A formattable object with colored backgrounds
}
\description{
Format factor columns with funky colored backgrounds
}
