% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_phyloseq.R
\name{remove_phyloseq}
\alias{remove_phyloseq}
\title{Remove a phyloseq object from a list_phyloseq}
\usage{
remove_phyloseq(x, name, verbose = TRUE)
}
\arguments{
\item{x}{(required) A list_phyloseq object.}

\item{name}{(character or integer) Name or index of the phyloseq object to remove.}
}
\value{
A new list_phyloseq object without the removed phyloseq
}
\description{
Remove a phyloseq object from a list_phyloseq
}
\examples{
lpq <- list_phyloseq(list(run1 = data_fungi, run2 = data_fungi_mini),
  verbose = FALSE
)
lpq2 <- remove_phyloseq(lpq, "run2", verbose = FALSE)
length(lpq2)
}
