% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtab_modification.R
\name{resolve_taxo_conflict}
\alias{resolve_taxo_conflict}
\title{Resolve taxonomic conflict in the tax_table of a phyloseq object}
\usage{
resolve_taxo_conflict(
  physeq,
  pattern_tax_ranks = NULL,
  method = c("consensus", "rel_majority", "abs_majority", "preference", "unanimity"),
  strict = FALSE,
  second_method = c("consensus", "rel_majority", "abs_majority", "preference",
    "unanimity"),
  nb_agree_threshold = 1,
  keep_tax_ranks = TRUE,
  new_names = NULL,
  preference_pattern = NULL,
  collapse_string = "/",
  replace_collapsed_rank_by_NA = FALSE
)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{pattern_tax_ranks}{A vector of pattern to aggregate taxonomic ranks.
For example "^Genus" stand for all taxonomic ranks (columns in tax_table
slot) starting by Genus}

\item{method}{One of "consensus", "rel_majority", "abs_majority",
"preference" or "unanimity". See details in the documentation of the
function \code{\link[MiscMetabar:resolve_vector_ranks]{MiscMetabar::resolve_vector_ranks()}}.}

\item{strict}{(logical, default FALSE). If TRUE, NA are considered as
informative in resolving conflict (i.e. NA are taking into account in vote).
See details for more informations.}

\item{second_method}{One of "consensus", "rel_majority", "abs_majority",
or "unanimity". Only used if method = "preference". See details.}

\item{nb_agree_threshold}{(Int, default 1) The minimum number of times a
value must arise to be selected using vote. If 2, we only kept
taxonomic value present at least 2 times in the vector.}

\item{keep_tax_ranks}{(logical, default TRUE) Do we keep the old taxonomic
ranks?}

\item{new_names}{A vector of new names for the taxonomic columns}

\item{preference_pattern}{A pattern to match the only column used as prefered
one if method = "preference".}

\item{collapse_string}{(default '/'). The character to collapse taxonomic names
when multiple assignment is done.}

\item{replace_collapsed_rank_by_NA}{(logical, default FALSE). If set to TRUE,
all multiple assignments (all taxonomic rank including the 'collapse_string'
parameter) are replaced by NA.}
}
\value{
A phyloseq object
}
\description{
Resolve taxonomic conflict in the tax_table of a phyloseq object
}
\examples{

data_fungi_mini_new <- assign_sintax(data_fungi_mini,
  ref_fasta = system.file("extdata", "mini_UNITE_fungi.fasta.gz",
    package = "MiscMetabar"
  ),
  behavior = "add_to_phyloseq"
)

resolve_taxo_conflict(data_fungi_mini_new, pattern_tax_ranks = c("^Genus\\\\."), method = "consensus", new_names = "Genus_consensus")@tax_table
resolve_taxo_conflict(data_fungi_mini_new, pattern_tax_ranks = c("^Genus\\\\."), method = "consensus", new_names = "Genus_consensus", replace_collapsed_rank_by_NA = TRUE)@tax_table

resolve_taxo_conflict(data_fungi_mini_new, pattern_tax_ranks = c("^Genus\\\\."), method = "preference", preference_pattern = ".x$")@tax_table
resolve_taxo_conflict(data_fungi_mini_new, pattern_tax_ranks = c("^Genus\\\\."), method = "abs_majority")@tax_table
resolve_taxo_conflict(data_fungi_mini_new, pattern_tax_ranks = c("^Genus\\\\."), method = "rel_majority")@tax_table
resolve_taxo_conflict(data_fungi_mini_new, pattern_tax_ranks = c("^Genus\\\\."), method = "unanimity")@tax_table

resolve_taxo_conflict(data_fungi_mini_new, pattern_tax_ranks = c("^Genus\\\\.", "^Family\\\\.", "^Species\\\\."), method = "consensus")@tax_table
}
\author{
Adrien Taudière
}
