% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxo_plot.R
\name{tc_bar}
\alias{tc_bar}
\title{Barchart of ratio to compare 2 taxonomic ranks}
\usage{
tc_bar(
  physeq,
  rank_1,
  rank_2,
  color_rank,
  point_size = 0.3,
  point_alpha = 0.3,
  merge_sample_by = NULL,
  log10trans = TRUE
)
}
\arguments{
\item{physeq}{(required): a \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{rank_1}{Define the first taxonomic rank as the number or the name
of the column in tax_table slot}

\item{rank_2}{Define the second taxonomic rank as the number or the name
of the column in tax_table slot}

\item{color_rank}{Define the taxonomic rank for color
as the number or the name of the column in tax_table slot}

\item{point_size}{Size of points}

\item{point_alpha}{Transparency of points}

\item{merge_sample_by}{a vector to determine
which samples to merge using \code{\link[=merge_samples2]{merge_samples2()}} function.
Need to be in \code{physeq@sam_data}}

\item{log10trans}{(Logical, default TRUE) Do the abundancy is log10
transformed?}
}
\value{
A ggplot2 object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Useful to compare taxonomy from two different source/db/algo side-by-side
}
\examples{

tc_bar(subset_taxa_pq(Glom_otu, taxa_sums(Glom_otu) > 5000), rank_1 = 5, rank_2 = 13, color_rank = 3)
\dontrun{
tc_bar(Glom_otu, rank_1 = 5, rank_2 = 13, color_rank = 3)
tc_bar(as_binary_otu_table(Glom_otu), rank_1 = 5, rank_2 = 13, color_rank = 3, log10trans = FALSE)
tc_bar(Glom_otu,
  rank_1 = "Genus",
  rank_2 = "Genus__eukaryome_Glomero",
  color_rank = "Family"
)
}
}
\author{
Adrien Taudière
}
