% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxo_plot.R
\name{tc_heatmap}
\alias{tc_heatmap}
\title{Heatmap of correspondence between two taxonomic ranks}
\usage{
tc_heatmap(
  physeq,
  rank_1,
  rank_2,
  low_color = "white",
  high_color = "steelblue",
  na_value = "grey80",
  merge_sample_by = NULL
)
}
\arguments{
\item{physeq}{(required) A \code{\link[phyloseq]{phyloseq-class}} object obtained
using the \code{phyloseq} package.}

\item{rank_1}{(character or integer) Define the first taxonomic rank as the
number or the name of the column in tax_table slot.}

\item{rank_2}{(character or integer) Define the second taxonomic rank as the
number or the name of the column in tax_table slot.}

\item{low_color}{(character, default "white") Color for low count cells.}

\item{high_color}{(character, default "steelblue") Color for high count cells.}

\item{na_value}{(character, default "grey80") Color for NA count cells.}

\item{merge_sample_by}{(character, default NULL) A vector to determine
which samples to merge using \code{\link[MiscMetabar:merge_samples2]{MiscMetabar::merge_samples2()}} function.
Need to be in \code{physeq@sam_data}.}
}
\value{
A ggplot2 object
}
\description{
\if{html}{\out{
<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
<img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
}}


Useful to compare taxonomy from two different source/db/algo side-by-side.
Each cell shows the number of taxa assigned to the combination of the two
ranks.
}
\examples{
tc_heatmap(
  subset_taxa_pq(Glom_otu, taxa_sums(Glom_otu) > 5000),
  rank_1 = "Family",
  rank_2 = "Family__eukaryome_Glomero"
)
\dontrun{
tc_heatmap(Glom_otu, rank_1 = 5, rank_2 = 13)
tc_heatmap(Glom_otu, rank_1 = "Genus", rank_2 = "Genus__eukaryome_Glomero",
  high_color = "darkred"
)
}
}
\author{
Adrien Taudière
}
